requirements.txt
streamlit
spacy
spacy-streamlit
pdfminer.six
gensim
rank_bm25
https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.0.0/de_core_news_sm-3.0.0.tar.gz
https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.0.0/de_core_news_lg-3.0.0.tar.gz